Javdoc for core
